예외 처리 핸들러 통합
